style: add responsive style AnnounceMessageBox
