+++ app/ui/view/molecule/announce-message-box/announce-massage-box.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import CompletedImage from '@/public/assets/completed-category.png';
+import CheckImage from '@/public/assets/check.svg';
+import Image from 'next/image';
+import { cn } from '@/app/utils/shadcn/utils';
+
+interface AnnounceMessageBoxProp {
+  message: string;
+}
+
+function AnnounceMessageBox({ message }: AnnounceMessageBoxProp) {
+  return (
+    <div className="relative flex justify-center items-center p-2 bg-yellow-5 min-h-fit">
+      <Image src={CompletedImage} width={1200} height={270} alt="category-fullfill-image" className="min-h-20" />
+      <div className="absolute">
+        <Image
+          className={cn('m-auto w-[30px]', 'md:w-[50px]')}
+          src={CheckImage}
+          width={50}
+          height={36}
+          alt="category-fullfill-image"
+        />
+        <p className={cn('zIndex-1 text-sm', 'md:text-lg')}>{message}</p>
+      </div>
+    </div>
+  );
+}
+
+export default AnnounceMessageBox;
